refactor: [#1298] move functions to container mod
