+++ src/servers/apis/server.rs
@@ -298,7 +298,8 @@ mod tests {
     use torrust_server_lib::registar::Registar;
     use torrust_tracker_test_helpers::configuration::ephemeral_public;
 
-    use crate::bootstrap::app::{initialize_global_services, initialize_http_api_container};
+    use crate::bootstrap::app::initialize_global_services;
+    use crate::container::initialize_http_api_container;
     use crate::servers::apis::server::{ApiServer, Launcher};
 
     #[tokio::test]
